pyproject.toml
[project]
name = "navai-stt-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "huggingface-hub>=0.34.4",
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "transformers>=4.56.0",
    "trl>=0.22.1",
    "wandb>=0.21.3",
]
